+++ firebase-ai/src/testUtil/java/com/google/firebase/ai/JavaCompileTests.java
@@ -21,9 +21,11 @@ import androidx.annotation.Nullable;
 import com.google.common.util.concurrent.ListenableFuture;
 import com.google.firebase.ai.FirebaseAI;
 import com.google.firebase.ai.GenerativeModel;
+import com.google.firebase.ai.ImagenModel;
 import com.google.firebase.ai.LiveGenerativeModel;
 import com.google.firebase.ai.java.ChatFutures;
 import com.google.firebase.ai.java.GenerativeModelFutures;
+import com.google.firebase.ai.java.ImagenModelFutures;
 import com.google.firebase.ai.java.LiveModelFutures;
 import com.google.firebase.ai.java.LiveSessionFutures;
 import com.google.firebase.ai.type.BlockReason;
@@ -33,6 +35,7 @@ import com.google.firebase.ai.type.CitationMetadata;
 import com.google.firebase.ai.type.Content;
 import com.google.firebase.ai.type.ContentModality;
 import com.google.firebase.ai.type.CountTokensResponse;
+import com.google.firebase.ai.type.Dimensions;
 import com.google.firebase.ai.type.FileDataPart;
 import com.google.firebase.ai.type.FinishReason;
 import com.google.firebase.ai.type.FunctionCallPart;
@@ -43,6 +46,11 @@ import com.google.firebase.ai.type.HarmCategory;
 import com.google.firebase.ai.type.HarmProbability;
 import com.google.firebase.ai.type.HarmSeverity;
 import com.google.firebase.ai.type.ImagePart;
+import com.google.firebase.ai.type.ImagenBackgroundMask;
+import com.google.firebase.ai.type.ImagenEditMode;
+import com.google.firebase.ai.type.ImagenEditingConfig;
+import com.google.firebase.ai.type.ImagenInlineImage;
+import com.google.firebase.ai.type.ImagenMaskReference;
 import com.google.firebase.ai.type.InlineDataPart;
 import com.google.firebase.ai.type.LiveGenerationConfig;
 import com.google.firebase.ai.type.LiveServerContent;
@@ -65,6 +73,7 @@ import com.google.firebase.ai.type.Voice;
 import com.google.firebase.concurrent.FirebaseExecutors;
 import java.util.ArrayList;
 import java.util.Calendar;
+import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.Executor;
@@ -139,6 +148,17 @@ public class JavaCompileTests {
         .setResponseModality(ResponseModality.AUDIO)
         .setSpeechConfig(new SpeechConfig(new Voice("AOEDE")))
         .build();
+  }
+
+  private void testImagen() {
+    ImagenModel modelSuspend = FirebaseAI.getInstance().imagenModel("");
+    ImagenModelFutures model = ImagenModelFutures.from(modelSuspend);
+    model.editImage(
+        Collections.singletonList(new ImagenBackgroundMask()),
+        "",
+        new ImagenEditingConfig(ImagenEditMode.OUTPAINT, 25));
+    ImagenMaskReference.generateMaskAndPadForOutpainting(
+        new ImagenInlineImage(new byte[0], ""), new Dimensions(0, 0));
   }
 
   private void testFutures(GenerativeModelFutures futures) throws Exception {